+++ lib/cloudinary/api.rb
@@ -649,7 +649,7 @@ class Cloudinary::Api
   #
   # @see https://cloudinary.com/documentation/admin_api#update_an_upload_preset
   def self.update_upload_preset(name, options={})
-    params = Cloudinary::Uploader.build_upload_params(options)
+    params = Cloudinary::Uploader.build_upload_params(options, true)
     call_api(:put, "upload_presets/#{name}", params.merge(only(options, :unsigned, :disallow_public_id, :live)), options)
   end
 
@@ -664,7 +664,7 @@ class Cloudinary::Api
   #
   # @see https://cloudinary.com/documentation/admin_api#create_an_upload_preset
   def self.create_upload_preset(options={})
-    params = Cloudinary::Uploader.build_upload_params(options)
+    params = Cloudinary::Uploader.build_upload_params(options, true)
     call_api(:post, "upload_presets", params.merge(only(options, :name, :unsigned, :disallow_public_id, :live)), options)
   end
 
@@ -1222,6 +1222,25 @@ class Cloudinary::Api
     call_metadata_rules_api(:delete, uri, {}, options)
   end
 
+  # Analyzes an asset with the requested analysis type.
+  #
+  # @param [Object] input_type    The type of input for the asset to analyze ('uri').
+  # @param [Object] analysis_type The type of analysis to run ('google_tagging', 'captioning', 'fashion').
+  # @param [Hash] options         The optional parameters.
+  #
+  # @return [Cloudinary::Api::Response]
+  #
+  # @raise [Cloudinary::Api::Error]
+  def self.analyze(input_type, analysis_type, options = {})
+    api_uri = ["analysis", "analyze", input_type]
+    params = only(options, :uri, :parameters)
+    params["analysis_type"] = analysis_type
+
+    options[:api_version] = 'v2'
+
+    call_api(:post, api_uri, params, options)
+  end
+
   protected
 
   # Execute a call api for input params.
@@ -1236,13 +1255,14 @@ class Cloudinary::Api
     api_key     = options[:api_key] || Cloudinary.config.api_key
     api_secret  = options[:api_secret] || Cloudinary.config.api_secret
     oauth_token = options[:oauth_token] || Cloudinary.config.oauth_token
+    api_version = options[:api_version] || Cloudinary.config.api_version || 'v1_1'
 
     validate_authorization(api_key, api_secret, oauth_token)
 
     auth = { :key => api_key, :secret => api_secret, :oauth_token => oauth_token }
 
     call_cloudinary_api(method, uri, auth, params, options) do |cloudinary, inner_uri|
-      [cloudinary, 'v1_1', cloud_name, inner_uri]
+      [cloudinary, api_version, cloud_name, inner_uri]
     end
   end
 
@@ -1254,7 +1274,7 @@ class Cloudinary::Api
     return Cloudinary::Utils.json_decode(response.body)
   rescue => e
     # Error is parsing json
-    raise GeneralError.new("Error parsing server response (#{response.code}) - #{response.body}. Got - #{e}")
+    raise GeneralError.new("Error parsing server response (#{response.status}) - #{response.body}. Got - #{e}")
   end
 
   # Protected function that assists with performing an API call to the metadata_fields part of the Admin API.