+++ spec/image_spec.rb
@@ -25,16 +25,17 @@ describe 'Responsive breakpoints' do
 
   before :each do
     Cloudinary.reset_config
+    Cloudinary.config.analytics = false
   end
 
   let(:cloud_name) {COMMON_TRANS[:cloud_name]}
-  let(:root_path) {"http://res.cloudinary.com/#{cloud_name}"}
+  let(:root_path) {"https://res.cloudinary.com/#{cloud_name}"}
   let(:upload_path) {"#{root_path}/image/upload"}
 
   let(:options) {COMMON_TRANS}
-  # let(:helper) {helper_class.new}
+
   let(:helper) {
-    ActionView::Base.new(ActionView::LookupContext.new([]))
+    ActionView::Base.new(ActionView::LookupContext.new([]), {}, nil)
   }
   let(:test_tag) {TestTag.new(helper.cl_image_tag('sample.jpg', options))}
   describe 'srcset' do