+++ spec/video_tag_spec.rb
@@ -27,6 +27,7 @@ describe CloudinaryHelper do
       config.secure_distribution = nil
       config.private_cdn         = false
       config.secure              = false
+      config.analytics           = false
       config.cname               = nil
       config.cdn_subdomain       = false
       config.api_key             = "1234"
@@ -35,7 +36,7 @@ describe CloudinaryHelper do
   end
 
   let(:helper) {
-    ActionView::Base.new(ActionView::LookupContext.new([]))
+    ActionView::Base.new(ActionView::LookupContext.new([]), {}, nil)
   }
   let(:root_path) { "http://res.cloudinary.com/#{DUMMY_CLOUD}" }
   let(:upload_path) { "#{root_path}/video/upload" }